refactor: 컨트롤러 테스트 네이밍 수정
