+++ src/test/java/kitchenpos/ui/ControllerTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.controller;
+package kitchenpos.ui;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
